+++ back/src/main/java/bsu/practice/service/servlet/BlankServlet.java
