+++ back/src/main/java/bsu/practice/service/servlet/PageServlet.java
@@ -1,4 +1,4 @@
-package bsu.practice;
+package bsu.practice.service.servlet;
 
 import java.io.IOException;
 
@@ -9,12 +9,12 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet(name="status", urlPatterns = "/status")
-public class StatusServlet extends HttpServlet {
+@WebServlet(name="page", urlPatterns = "/page")
+public class PageServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        RequestDispatcher rd = req.getRequestDispatcher("WEB-INF/status.jsp");
+        RequestDispatcher rd = req.getRequestDispatcher("WEB-INF/page.html");
         if (rd != null) {
             rd.forward(req, resp);
         }
