+++ back/src/main/java/bsu/practice/service/servlet/Test2Servlet.java
@@ -1,4 +1,4 @@
-package bsu.practice;
+package bsu.practice.service.servlet;
 
 import java.io.IOException;
